+++ include/list.h
@@ -19,7 +19,7 @@
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307
  *  USA
  *
- *  $Id: list.h,v 7.17 2003/04/16 09:46:58 michael Exp $
+ *  $Id: list.h,v 7.18 2003/05/22 17:53:23 joshk Exp $
  */
 
 #ifndef INCLUDED_list_h
@@ -33,8 +33,8 @@ struct Class;
 struct User;
 
 extern void initUser(void);
-extern void count_user_memory(int *count, int *user_memory_used);
-extern void count_links_memory(int *count, int *links_memory_used);
+extern void count_user_memory(int *count, unsigned long *user_memory_used);
+extern void count_links_memory(int *count, unsigned long *links_memory_used);
 extern void free_user(struct User *, struct Client *);
 extern dlink_node *make_dlink_node(void);
 extern void free_dlink_node(dlink_node *lp);
